Reorganized code and files.
