+++ PolygonChecker/rect.h
@@ -1,0 +1,5 @@
+#pragma once
+
+double distanceSquared(int x1, int y1, int x2, int y2);
+int isRectangle(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4);
+void checkRectangle();
